Using ECM vs CJS in handler. Updated default thresholds for AlarmClassifications to 95 and 90 respectively. Updated README.md to reflect new default values. Removed various unused files.
