+++ handlers/.gitignore
@@ -1,0 +1,5 @@
+node_modules
+*.d.*
+*.map
+*.mjs
+src/**/*.js
