+++ handlers/package.json
@@ -30,7 +30,7 @@
     "@aws-sdk/s3-request-presigner": "^3.600.0",
     "@aws-sdk/util-dynamodb": "^3.602.0",
     "@nr1e/commons": "0.0.2-alpha.9",
-    "@nr1e/logging": "^0.0.8",
+    "@nr1e/logging": "^1.1.0",
     "@smithy/protocol-http": "^4.0.2",
     "@smithy/signature-v4": "^3.1.1",
     "@types/js-yaml": "^4.0.9",
