+++ handlers/src/alb-modules.mts
@@ -1,34 +1,32 @@
 import {
   ElasticLoadBalancingV2Client,
   DescribeTagsCommand,
+  DescribeLoadBalancersCommand,
 } from '@aws-sdk/client-elastic-load-balancing-v2';
 import * as logging from '@nr1e/logging';
-import {AlarmProps, Tag} from './types';
-import {AlarmClassification, ValidTargetGroupEvent} from './enums';
+import {AlarmProps, Tag} from './types.mjs';
+import {AlarmClassification, ValidAlbEvent} from './enums.mjs';
 import {
   createOrUpdateCWAlarm,
   getCWAlarmsForInstance,
   deleteCWAlarm,
-} from './alarm-tools';
+} from './alarm-tools.mjs';
 
-const log: logging.Logger = logging.getRootLogger();
+const log: logging.Logger = logging.getLogger('alb-modules');
 const elbClient: ElasticLoadBalancingV2Client =
   new ElasticLoadBalancingV2Client({});
-
 const defaultThresholds: {[key in AlarmClassification]: number} = {
-  [AlarmClassification.Critical]: 90,
-  [AlarmClassification.Warning]: 80,
+  [AlarmClassification.Critical]: 15000,
+  [AlarmClassification.Warning]: 10000,
 };
-
 const metricConfigs = [
-  {metricName: 'TargetResponseTime', namespace: 'AWS/ApplicationELB'},
-  {metricName: 'RequestCountPerTarget', namespace: 'AWS/ApplicationELB'},
-  {metricName: 'HTTPCode_Target_4XX_Count', namespace: 'AWS/ApplicationELB'},
-  {metricName: 'HTTPCode_Target_5XX_Count', namespace: 'AWS/ApplicationELB'},
+  {metricName: 'RequestCount', namespace: 'AWS/ApplicationELB'},
+  {metricName: 'HTTPCode_ELB_4XX_Count', namespace: 'AWS/ApplicationELB'},
+  {metricName: 'HTTPCode_ELB_5XX_Count', namespace: 'AWS/ApplicationELB'},
 ];
 
 async function getAlarmConfig(
-  targetGroupName: string,
+  loadBalancerName: string,
   type: AlarmClassification,
   metricName: string
 ): Promise<{
@@ -42,18 +40,18 @@ async function getAlarmConfig(
   const durationPeriodsKey = `autoalarm:${metricName}-percent-duration-periods`;
 
   return {
-    alarmName: `AutoAlarm-TargetGroup-${targetGroupName}-${type}-${metricName}`,
+    alarmName: `AutoAlarm-ALB-${loadBalancerName}-${type}-${metricName}`,
     thresholdKey,
     durationTimeKey,
     durationPeriodsKey,
   };
 }
 
-export async function fetchTargetGroupTags(
-  targetGroupArn: string
-): Promise<Tag> {
+export async function fetchALBTags(loadBalancerArn: string): Promise<Tag> {
   try {
-    const command = new DescribeTagsCommand({ResourceArns: [targetGroupArn]});
+    const command = new DescribeTagsCommand({
+      ResourceArns: [loadBalancerArn],
+    });
     const response = await elbClient.send(command);
     const tags: Tag = {};
 
@@ -67,33 +65,33 @@ export async function fetchTargetGroupTags(
 
     log
       .info()
-      .str('targetGroupArn', targetGroupArn)
+      .str('loadBalancerArn', loadBalancerArn)
       .str('tags', JSON.stringify(tags))
-      .msg('Fetched Target Group tags');
+      .msg('Fetched ALB tags');
 
     return tags;
   } catch (error) {
     log
       .error()
       .err(error)
-      .str('targetGroupArn', targetGroupArn)
-      .msg('Error fetching Target Group tags');
+      .str('loadBalancerArn', loadBalancerArn)
+      .msg('Error fetching ALB tags');
     return {};
   }
 }
 
-async function checkAndManageTargetGroupStatusAlarms(
-  targetGroupName: string,
+async function checkAndManageALBStatusAlarms(
+  loadBalancerName: string,
   tags: Tag
 ) {
   if (tags['autoalarm:disabled'] === 'true') {
     const activeAutoAlarms: string[] = await getCWAlarmsForInstance(
-      'TargetGroup',
-      targetGroupName
+      'ALB',
+      loadBalancerName
     );
     await Promise.all(
       activeAutoAlarms.map(alarmName =>
-        deleteCWAlarm(alarmName, targetGroupName)
+        deleteCWAlarm(alarmName, loadBalancerName)
       )
     );
     log.info().msg('Status check alarm creation skipped due to tag settings.');
@@ -103,7 +101,7 @@ async function checkAndManageTargetGroupStatusAlarms(
       for (const classification of Object.values(AlarmClassification)) {
         const {alarmName, thresholdKey, durationTimeKey, durationPeriodsKey} =
           await getAlarmConfig(
-            targetGroupName,
+            loadBalancerName,
             classification as AlarmClassification,
             metricName
           );
@@ -114,12 +112,12 @@ async function checkAndManageTargetGroupStatusAlarms(
           namespace: namespace,
           evaluationPeriods: 5,
           metricName: metricName,
-          dimensions: [{Name: 'TargetGroup', Value: targetGroupName}],
+          dimensions: [{Name: 'LoadBalancer', Value: loadBalancerName}],
         };
 
         await createOrUpdateCWAlarm(
           alarmName,
-          targetGroupName,
+          loadBalancerName,
           alarmProps,
           tags,
           // @ts-ignore
@@ -132,51 +130,48 @@ async function checkAndManageTargetGroupStatusAlarms(
   }
 }
 
-export async function manageTargetGroupAlarms(
-  targetGroupName: string,
+export async function manageALBAlarms(
+  loadBalancerName: string,
   tags: Tag
 ): Promise<void> {
-  await checkAndManageTargetGroupStatusAlarms(targetGroupName, tags);
+  await checkAndManageALBStatusAlarms(loadBalancerName, tags);
 }
 
-export async function manageInactiveTargetGroupAlarms(targetGroupName: string) {
+export async function manageInactiveALBAlarms(loadBalancerName: string) {
   try {
     const activeAutoAlarms: string[] = await getCWAlarmsForInstance(
-      'TargetGroup',
-      targetGroupName
+      'ALB',
+      loadBalancerName
     );
     await Promise.all(
       activeAutoAlarms.map(alarmName =>
-        deleteCWAlarm(alarmName, targetGroupName)
+        deleteCWAlarm(alarmName, loadBalancerName)
       )
     );
   } catch (e) {
-    log.error().err(e).msg(`Error deleting Target Group alarms: ${e}`);
-    throw new Error(`Error deleting Target Group alarms: ${e}`);
+    log.error().err(e).msg(`Error deleting ALB alarms: ${e}`);
+    throw new Error(`Error deleting ALB alarms: ${e}`);
   }
 }
 
-export async function getTargetGroupEvent(event: any): Promise<{
-  targetGroupArn: string;
-  eventName: ValidTargetGroupEvent;
-  tags: Tag;
-}> {
-  const targetGroupArn =
-    event.detail.responseElements?.targetGroups[0]?.targetGroupArn;
-  const eventName = event.detail.eventName as ValidTargetGroupEvent;
-
+export async function getAlbEvent(
+  event: any
+): Promise<{loadBalancerArn: string; eventName: ValidAlbEvent; tags: Tag}> {
+  const loadBalancerArn =
+    event.detail.responseElements?.loadBalancers[0]?.loadBalancerArn;
+  const eventName = event.detail.eventName as ValidAlbEvent;
   log
     .info()
-    .str('targetGroupArn', targetGroupArn)
+    .str('loadBalancerArn', loadBalancerArn)
     .str('eventName', eventName)
-    .msg('Processing Target Group event');
-  const tags = await fetchTargetGroupTags(targetGroupArn);
+    .msg('Processing ALB event');
+  const tags = await fetchALBTags(loadBalancerArn);
 
-  if (targetGroupArn && eventName === ValidTargetGroupEvent.Active) {
-    await manageTargetGroupAlarms(targetGroupArn, tags);
-  } else if (eventName === ValidTargetGroupEvent.Deleted) {
-    await manageInactiveTargetGroupAlarms(targetGroupArn);
+  if (loadBalancerArn && eventName === ValidAlbEvent.Active) {
+    await manageALBAlarms(loadBalancerArn, tags);
+  } else if (eventName === ValidAlbEvent.Deleted) {
+    await manageInactiveALBAlarms(loadBalancerArn);
   }
 
-  return {targetGroupArn, eventName, tags};
+  return {loadBalancerArn, eventName, tags};
 }
